+++ app/src/main/java/eu/kanade/presentation/anime/AnimeScreen.kt
@@ -73,7 +73,6 @@ import eu.kanade.presentation.util.formatEpisodeNumber
 import eu.kanade.tachiyomi.animesource.AnimeSource
 import eu.kanade.tachiyomi.animesource.model.FetchType
 import eu.kanade.tachiyomi.animesource.model.SAnime
-import eu.kanade.tachiyomi.data.download.DownloadProvider
 import eu.kanade.tachiyomi.data.download.model.Download
 import eu.kanade.tachiyomi.source.getNameForAnimeInfo
 import eu.kanade.tachiyomi.ui.anime.AnimeSeasonItem
@@ -88,6 +87,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 // <-- AM (NOW_PLAYING_INDICATOR)
+import mihon.app.di.appGraph
 import tachiyomi.core.common.util.lang.withIOContext
 import tachiyomi.domain.anime.model.Anime
 import tachiyomi.domain.anime.model.EpisodeViewMode
@@ -106,7 +106,6 @@ import tachiyomi.presentation.core.components.material.Scaffold
 import tachiyomi.presentation.core.i18n.stringResource
 import tachiyomi.presentation.core.util.shouldExpandFAB
 import tachiyomi.source.local.isLocal
-import uy.kohesive.injekt.injectLazy
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.Instant
@@ -1384,7 +1383,9 @@ private fun LazyGridScope.sharedEpisodeItems(
         span = { GridItemSpan(maxLineSpan) },
         // <-- AY
     ) { item ->
+        val context = LocalContext.current
         val haptic = LocalHapticFeedback.current
+        val downloadProvider = remember { context.appGraph.downloadProvider }
 
         when (item) {
             is EpisodeList.MissingCount -> {
@@ -1547,7 +1548,3 @@ private fun Modifier.ignorePadding(gridPadding: Int) = layout { measurable, cons
     }
 }
 // <-- AY
-
-// AM (FILE_SIZE) -->
-private val downloadProvider: DownloadProvider by injectLazy()
-// <-- AM (FILE_SIZE)