+++ app/src/main/java/mihon/domain/source/interactor/UpdateAnimeFromRemote.kt
@@ -1,5 +1,6 @@
 package mihon.domain.source.interactor
 
+import dev.zacsweers.metro.Inject
 import eu.kanade.domain.anime.interactor.SyncSeasonsWithSource
 import eu.kanade.domain.anime.model.hasCustomBackground
 import eu.kanade.domain.anime.model.hasCustomCover
@@ -29,6 +30,7 @@ import tachiyomi.source.local.LocalSource
 import tachiyomi.source.local.isLocal
 import kotlin.time.Clock
 
+@Inject
 class UpdateAnimeFromRemote(
     private val sourceManager: SourceManager,
     private val episodeRepository: EpisodeRepository,