+++ domain/src/main/java/tachiyomi/domain/anime/interactor/SetAnimeEpisodeFlags.kt
@@ -1,10 +1,12 @@
 package tachiyomi.domain.anime.interactor
 
+import dev.zacsweers.metro.Inject
 import tachiyomi.domain.anime.model.Anime
 import tachiyomi.domain.anime.model.AnimeUpdate
 import tachiyomi.domain.anime.model.EpisodeViewMode
 import tachiyomi.domain.anime.repository.AnimeRepository
 
+@Inject
 class SetAnimeEpisodeFlags(
     private val animeRepository: AnimeRepository,
 ) {