+++ CHANGELOG.md
@@ -28,7 +28,7 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 ## [v0.19.9.1] - 2026-08-12
 ### Other
 - Merge from aniyomi ([@Secozzi](https://github.com/Secozzi)) ([#191](https://github.com/quickdesh/Animiru/pull/191), [#200](https://github.com/quickdesh/Animiru/pull/200))
-- Merge from mihon and aniyomi ([@Secozzi](https://github.com/Secozzi)) ([#192](https://github.com/quickdesh/Animiru/pull/192))
+- Merge from mihon and aniyomi ([@Secozzi](https://github.com/Secozzi)) ([#192](https://github.com/quickdesh/Animiru/pull/192), [#207](https://github.com/quickdesh/Animiru/pull/207))
 - Rebrand app name and icon to Vidi
 
 ## [v0.19.8.1] - 2026-08-09